pfs-fuse: open s3 no need to fetch from remote

+++ pkg/fs/client/vfs_new/reader.go
@@ -124,7 +124,7 @@ func (fh *fileReader) Read(buf []byte, off uint64) (int, syscall.Errno) {
 			return 0, syscall.EBADF
 		}
 		// todo:: 不走缓存部分需要保持原来open-read模式，保证这部分性能
-		bytesRead, err = fh.fd.Read(buf, int64(off))
+		bytesRead, err = fh.fd.Read(buf, off)
 		if err != nil {
 			log.Errorf("ufs read err: %v", err)
 			return 0, syscall.EBADF
